% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyph.support-it.R
\name{hyph.support.it}
\alias{hyph.support.it}
\title{Language support for Italian}
\usage{
hyph.support.it()
}
\description{
This function adds support for Italian to the sylly package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular,
      a new set of hyphenation patterns is being added (see \code{\link{hyph.it}}).
To use the patterns with \code{\link[sylly:hyphen]{hyphen}}, simply use the abbreviation:
\itemize{
  \item {\code{"it"}} {--- Italian hyphenation patterns}
}
}
\seealso{
\code{\link[sylly:hyphen]{hyphen}},
   \code{\link[sylly:set.hyph.support]{set.hyph.support}}
}
